+++ simplex.cpp
